+++ pyads/__init__.py
@@ -114,4 +114,4 @@ from .constants import (
     ADSTRANS_SERVERONCHA,
 )
 
-__version__ = '3.2.2'
+__version__ = '3.3.0'